requirements.txt
fastapi
uvicorn
python-telegram-bot
pydantic
pydantic-settings